# apis/api/swagger/swagger.yaml
components:
...
    schemas:
        System:
            properties:
                _id:
                    type: string
                referencedParts:
                    items:
                        type: string
                    type: array
                description:
                    type: string
                    description: 'A full description of the system'
                summary:
                    type: string
                    description: 'Short summary of the system'
                name:
                    type: string
                    description: 'The name of the system'
            required:
                - _id
                - referencedParts
                - description
                - summary
                - name
            type: object
        Part:
            properties:
                _id:
                    type: string
                system:
                    items:
                        type: string
                    type: array
                    description: 'Systems the part can be used in'
                name:
                    type: string
                    description: 'The name of the part'
            required:
                - _id
                - system
                - name
            type: object
        Boat:
            properties:
                _id:
                    type: string
                hardware:
                    items:
                        properties:
                            parts:
                                items:
                                    $ref: '#/components/schemas/Part'
                                type: array
                            systems:
                                items:
                                    $ref: '#/components/schemas/System'
                                type: array
                        type: object
                    type: array
                dimensionData:
                    properties:
                        LP:
                            type: number
                            format: double
                            description: 'Shortest distance from headstay to the clew of the jib'
                        E:
                            type: number
                            format: double
                            description: 'Foot length of the mainsail'
                        P:
                            type: number
                            format: double
                            description: 'Luff length of the mainsail'
                        J2:
                            type: number
                            format: double
                            description: 'Base of staysail triangle'
                        J:
                            type: number
                            format: double
                            description: 'Base of the foretriangle measured from the front of the mast to the intersection of the forestay and deck'
                        I2:
                            type: number
                            format: double
                            description: 'Height of staysail halyard above deck'
                        I:
                            type: number
                            format: double
                            description: 'Height of the foretriangle measured from the top of the highest sheave to the sheerline'
                        BWL:
                            type: number
                            format: double
                            description: 'Beam Waterline - widest beam of boat at the waterline'
                        BMX:
                            type: number
                            format: double
                            description: 'Beam Maximum - width of the boat at the widest point'
                        DWL:
                            type: number
                            format: double
                            description: 'Design Waterline -theoretical waterline length of boat as opposed to LWL, which is actual waterline length'
                        LWL:
                            type: number
                            format: double
                            description: 'Length Waterline - length of waterline of the boat'
                        LOA:
                            type: number
                            format: double
                            description: 'Length Overall - overall tip-to-tip length of the boat'
                    required:
                        - LP
                        - E
                        - P
                        - J2
                        - J
                        - I2
                        - I
                        - BWL
                        - BMX
                        - DWL
                        - LWL
                        - LOA
                    type: object
                model:
                    type: string
                    description: 'The model of the boat'
                brand:
                    type: string
                    description: 'The brand of the boat'
                owner:
                    type: string
            required:
                - _id
                - model
                - brand
                - owner
            type: object
        Pick_Boat.Exclude_keyofBoat._id:
            properties:
                owner:
                    type: string
                brand:
                    type: string
                    description: 'The brand of the boat'
                model:
                    type: string
                    description: 'The model of the boat'
                dimensionData:
                    properties:
                        LP:
                            type: number
                            format: double
                            description: 'Shortest distance from headstay to the clew of the jib'
                        E:
                            type: number
                            format: double
                            description: 'Foot length of the mainsail'
                        P:
                            type: number
                            format: double
                            description: 'Luff length of the mainsail'
                        J2:
                            type: number
                            format: double
                            description: 'Base of staysail triangle'
                        J:
                            type: number
                            format: double
                            description: 'Base of the foretriangle measured from the front of the mast to the intersection of the forestay and deck'
                        I2:
                            type: number
                            format: double
                            description: 'Height of staysail halyard above deck'
                        I:
                            type: number
                            format: double
                            description: 'Height of the foretriangle measured from the top of the highest sheave to the sheerline'
                        BWL:
                            type: number
                            format: double
                            description: 'Beam Waterline - widest beam of boat at the waterline'
                        BMX:
                            type: number
                            format: double
                            description: 'Beam Maximum - width of the boat at the widest point'
                        DWL:
                            type: number
                            format: double
                            description: 'Design Waterline -theoretical waterline length of boat as opposed to LWL, which is actual waterline length'
                        LWL:
                            type: number
                            format: double
                            description: 'Length Waterline - length of waterline of the boat'
                        LOA:
                            type: number
                            format: double
                            description: 'Length Overall - overall tip-to-tip length of the boat'
                    required:
                        - LP
                        - E
                        - P
                        - J2
                        - J
                        - I2
                        - I
                        - BWL
                        - BMX
                        - DWL
                        - LWL
                        - LOA
                    type: object
                hardware:
                    items:
                        properties:
                            parts:
                                items:
                                    $ref: '#/components/schemas/Part'
                                type: array
                            systems:
                                items:
                                    $ref: '#/components/schemas/System'
                                type: array
                        type: object
                    type: array
            required:
                - owner
                - brand
                - model
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        OmitOverride_Boat._id_:
            $ref: '#/components/schemas/Pick_Boat.Exclude_keyofBoat._id'
        BoatCreate:
            $ref: '#/components/schemas/OmitOverride_Boat._id_'
        Pick_Part.Exclude_keyofPart._id:
            properties:
                name:
                    type: string
                    description: 'The name of the part'
                system:
                    items:
                        type: string
                    type: array
                    description: 'Systems the part can be used in'
            required:
                - name
                - system
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        OmitOverride_Part._id_:
            $ref: '#/components/schemas/Pick_Part.Exclude_keyofPart._id'
        PartCreate:
            $ref: '#/components/schemas/OmitOverride_Part._id_'
...
    securitySchemes:
        jwt:
            type: apiKey
            name: authorization
            in: header
info:
    title: '@helium/ts-tsoa-api'
    version: 0.0.2
    description: 'Helium API with TSOA generator and openAPI sdk generator'
    contact: {}
openapi: 3.0.0
paths:
    /boat/find/all:
        post:
            operationId: AdminFindboat
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Boat'
                                type: array
            tags:
                - boat
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /boat/find:
        post:
            operationId: Findboats
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Boat'
                                type: array
            tags:
                - boat
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/boat/{boatId}':
        get:
            operationId: Getboat
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Boat'
            tags:
                - boat
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: boatId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updateboat
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Boat'
            tags:
                - boat
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: boatId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BoatCreate'
        delete:
            operationId: Deleteboat
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - boat
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: boatId
                    required: true
                    schema:
                        type: string
    /boat:
        post:
            operationId: Createboat
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Boat'
            tags:
                - boat
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BoatCreate'
    /part/find/all:
        post:
            operationId: AdminFindpart
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Part'
                                type: array
            tags:
                - part
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /part/find:
        post:
            operationId: Findparts
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Part'
                                type: array
            tags:
                - part
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/part/{partId}':
        get:
            operationId: Getpart
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Part'
            tags:
                - part
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: partId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updatepart
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Part'
            tags:
                - part
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: partId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PartCreate'
        delete:
            operationId: Deletepart
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - part
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: partId
                    required: true
                    schema:
                        type: string
    /part:
        post:
            operationId: Createpart
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Part'
            tags:
                - part
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PartCreate'
...
servers:
    -
        url: 'http://localhost:3002/v1'